Remove duplicated test; move mock classes to mocks.dart
