+++ packages/patrol_cli/test/commands/patrol_build_android_print_paths_test.dart
@@ -1,11 +1,6 @@
 import 'package:mocktail/mocktail.dart';
 import 'package:path/path.dart' show join;
 import 'package:patrol_cli/src/commands/build_android.dart';
-import 'package:patrol_cli/src/compatibility_checker/compatibility_checker.dart';
-import 'package:patrol_cli/src/dart_defines_reader.dart';
-import 'package:patrol_cli/src/pubspec_reader.dart';
-import 'package:patrol_cli/src/test_bundler.dart';
-import 'package:patrol_cli/src/test_finder.dart';
 import 'package:test/test.dart';
 
 import '../src/mocks.dart';
@@ -244,8 +239,9 @@ void main() {
         // Currently testing that null and empty string behaviors are documented through tests
 
         command.printApkPaths(buildMode: 'debug');
-        final nullFlavorOutput =
-            verify(() => mockLogger.info(captureAny())).captured;
+        final nullFlavorOutput = verify(
+          () => mockLogger.info(captureAny()),
+        ).captured;
 
         // Reset mock
         reset(mockLogger);
@@ -269,14 +265,3 @@ void main() {
     });
   });
 }
-
-// Create minimal mock classes for dependencies that don't need to be tested
-class MockTestFinder extends Mock implements TestFinder {}
-
-class MockTestBundler extends Mock implements TestBundler {}
-
-class MockDartDefinesReader extends Mock implements DartDefinesReader {}
-
-class MockPubspecReader extends Mock implements PubspecReader {}
-
-class MockCompatibilityChecker extends Mock implements CompatibilityChecker {}
